Refactor icon.scss to be simpler
- This commit moves the `@if map-has-key($state-list, $state)` check out of the
inline call and into the method for getting the call. This makes the Sass tidier
and easier to understand

- This commit also adds styling to the :after element to buttons so that the
SVG/PNG for the background image for the hoverstate is already loaded onto the
page so that there is no Flash Of Invisible Content (FOIC) if the user hovers
over the button while the browser fetches the resource.