Replace CSS cross with o-icons cross.
This commit replaces the CSS cross for the close button with one from o-icons.

This commit does not introduce a dependency on o-icons, rather it gets the cross
from the image service. This is so that this change doesn't have to be a
breaking change, and so that this version can remain compatible with older
versions of components which use older versions of o-[ft]-icons (because they
use the icon font still)

Some future work should be done to use the mixin provided by o-icons to set the
background image of the close button instead of doing it directly in the Sass here.

This commit:
- Removes the internal span (o-overlay__close-icon) from the close button as
it's no longer needed
- Adds a background image to the overlay