Add dependency on o-icons
This commit replaces the code duplicated from o-icons with the call
to the actual o-icons mixin.

The introduction of a dependency on o-icons _may_ break builds
if o-icons is also being requested at a different build elsewhere.

o-icons has been included at v4 or 5 in order to broaden the number
of resolvable versions of o-icons wider for easier integration.