Remove link focus styles and colors dep
So...

Adding blanket focus styles for links that change the background-color
and color properties has some weird side effects on things that have
already defined or part defined their own focus styles. These side
effects are things like:

```css
// In o-normalise
a:focus {
  color: red;
  background-color:black
}

// In some other component or bit of the page
.o-component a:focus {
  color:black
}
```

This results in the focus state for the `o-component a` being black on
black.

So, to avoid this problem, we're just going to set the outline, which
shouldn't introduce any dire user experience problems as browsers
already use this to style focus states as a default.

This commit also removes the o-colors dependency as we can go without it
and so avoid a breaking change.