Write SteppedProgress JavaScript (#10)
This adds some fairly minimal JavaScript to allow for users to interact
with stepped progress. The main use case is progressing the steps,
marking the current step as complete and the next step as current.

It's also possible to directly manipulate each step which caters for
moving backwards through the stepped progress. We can provide
convenience methods for this use case later if we need to.

The best way to see how this works is probably to look at the README
changes first. There's also full JSDoc for all public and private
methods which should hopefully make their purpose clear.