Prevent flash of the drawer on mobile.
The core experience of o-header-services is a little broken.
- There's a flash of the drawer as the JS loads on each page.
- The drop down is detatched from the menu item when items are
  on multiple lines (due to the drawer not initalising).
- The scroll arrows show even though they don't work.

This PR restores the core behaviour of the previous major
version (v2):

- Hide primary nav items for the core experience when they
  should be behind the drawer. There should be a fallback
provided using the hamburger's anchor. This behaviour is
consistent with the demos and previous major (v2), I've
reinstated our previous recomendations to the README (https://github.com/Financial-Times/o-header-services/tree/v2.3.4#core-experience-of-the-drawer).
- Hide the scroll arrows within the secondary nav for the core
  experience.
- Remove some redundant CSS along the way.