Document deprecated & potential deprecatation events
error.livefyreJs = Being deprecated because livefyre is being removed

tracking.socialMention = Being deprecated because its not possible to
mention someone in a comment with Coral Talk

widget.ready = This feels like it reflects
how the component was built rather than a significant moment.
Could possibly be replaced by oher component eventsMore investigation needed

widget.load = This feels like it reflects
how the component was built rather than a significant moment.
Could possibly be replaced by oher component eventsMore investigation needed

tracking.shareComment = Coral Talk only gives a simple share link and no
social options, this is something we are happy with. This event will
become a lot less meaningful if it only tells you the button was
clicked. It might stay for that reason only but will likely the
interface for it will change a lot