Fix drawer/dropdown race condition.
The Dropdown checks if the drawer is enabled by looking for a CSS
class. The Drawer adds the CSS class when it should be enabled.
There was a race condition, where the Drawer would add the class
on resize with a debounced event listener after the Dropdown
has checked for the class.

This PR adds a debounce to the Dropdown resize for performance
purposes, but a race condition could still exist. So the Drawer
is passed to the Dropdown so the Dropdown may check the status
of the Drawer directly without an implicit dependency.

There is a "flash" during the transition as the nav item still
animates -- this is a seperate issue.
https://github.com/Financial-Times/o-header-services/issues/107