Rename events to work with event listeners
Previously o-comments wasn't mentioned in the event naming because I had
only considered events being accessed via the `.on` method.

However events need to be accessible via event listeners as well and in
this case it would be nicer to include a namespacing.

This also includes some updates based on what events are accessible in
Talk v5