Render comments and login in parallel
Rather than promise changing the fetch of the JWT and the rendering of
comments making them happen in parallel.

This is now possible that we know how to login the user after the
component has rendered.

I hope that this will reduce the complexity in .renderComments and help
us split it into granular tasks.