Reduce the amount of events that are emitted
We are dispatching lots of events in order to handle event mapping from
Coral Talk to o-comments. This was originally done as it was taken from
the Talk docs but now that our code has matured we have better ways of
handling this and can reduce the complexity caused by firing and
listening for events.