Remove the `.on` interface
We are going to switch to just emitting events and suggesting users
listen to events via `addEventListener`.

For tracking we will also emit some common oTracking events that can be
picked up by oTracking should it be on the page. This will help prevent
differences between ft.com and the app.