Update event names to match Coral Talk new events
The event names now match the event names that Coral consider to be
their public event names.

This is the minimum update to get the basic events changed and matches
the behaviour before they changed the event names.

We can still make further improvements to this to track the
auto-moderated comments and to make sure they are not tracked as
successful comments.