Destroy display name prompt overlay when closed
With the Origami major update, we started getting an error (o-overlay with id "displayName" already exists) when trying to reopen the display name prompt after once it has been closed.

Because there is no event named `oOverlay.close`, we are now listening to `oLayers.close` and when it is dispatched, we destroy the overlay.