Add a display name option
This will allow the display name to be passed into o-comments as an
option.

This will allow the parent application to set the display name in
advance should it exist. This will be useful for Longroom room where we
want the display name to be set in the user doesn't already have a
commenting account.

```
<div class="o-comments"
	id="{element-id}"
	data-o-component="o-comments"
	data-o-comments-display-name="Test display name">
</div>
```