Add a mixin to allow for output of a custom theme
This mixin is based on the way we allow customisation for o-labels. It
outputs a CSS selector which is an appended o-banner class, which
contains all of the styles for the custom banner.

I'm having to provide a lot of defaults in this so that a user doesn't
need to specify every property to get a reasonable customised banner.
There may be a nicer way to do this.