Update timestamp alignment for related assassination teasers.
Live timestamps are usually ordered so they move to the top of a
teaser, above the heading and tag. In the assassination related
section, where multiple teasers (live and not-live) are displayed
in a row, the live timestamp should remain in place to align with
other timestamps.

Note the related assassination teaser collection is in a bad way.
Styles a split between o-teaser-collection and next-front-page.
It also does not conform to the current Origami specification because
it uses `o-teaser` selectors, to modify the styles of a separate
component.
https://github.com/Financial-Times/next-front-page/blob/309b844b5c25fc99fcc9cdab0152481237addf57/client/components/teaser-collections/_assassination.scss#L67