Add editorial blockquote.
This commit is a big one which was difficult to break up:

Updates the README with a Sass section, contact section, etc.

Renames private Sass placeholders to remove BEM syntax as they
no longer map directly to a class of the same name.

Deprecates `oQuoteBase` and `oQuoteBaseCite`. Users are able to
use `oQuoteStandard` and `oQuoteStandardCite` alone, or the new
editorial variant equivalents, without remembering to also include
the base mixins.

The base o-quote styles are stripped back so we don't have to
undo these styles in the future when we add other variants
such as a pull quote -- I'm not quite sure why base styles
exist at all, they don't seem to be used.

Maintains the current standard style for use in non-editorial
contexts, this may be a candidate for deprecation in the near future.

Adds a new `o-quote--editorial` modifier and associated Mixins.
Going forward this should be used in article content over the
standard version which is currently used on ft.com:
https://github.com/Financial-Times/n-content-body/blob/c703b86dd04561f240b2e6cb08f0472eba9566a2/scss/_blockquotes.scss#L5