Fix external links focus outline not showing
Make sure that when a user tabs to the link in the footer they know where they are by showing a focus outline.

https://www.w3.org/WAI/WCAG21/Understanding/focus-visible.html

The external link contains an override to the overflow. So needs to be after the regular matrix mixin.

Jira: CPP-115

DAC ID: DAC_Missing_Visible_Focus_Indicator_Issue1