Show comment action button icons on all screens
We used to hide these icons on smaller screen because there wasn't space
for them. In a recent update Coral have changed their behaviour on these
buttons and they now hide the text on smaller screens and only the icon
is visable so we need to stop hiding these icons otherwise nothing is
visable.