Replace paid post labels with partner content styles.
The commercial design team created new teaser designs to replace
"paid posts" with "partner content". The ads team worked with
the commercial design team to refine requirements and rollout
the new design to ft.com with overrides:
https://github.com/Financial-Times/n-native-ads/blob/0fa02c2a685085e9a3e183672569431670eb260a/main.scss

However neither the product design team or Origami were not
involved and so the design strikes out from other teasers.
E.g. the "partner content" label no longer looks like a label
but differs from tags in that it is not a link and is uppercase.

This commit brings the overrides into Origami to:
- make maintence easier
- reduce css bundle size
- roll the changes out everywhere, including the app
- update demos and document the change, so we have a
solid base for future updates by the product design team

Some limited changes have been made from the ads overrides:
- the hover state isn't oxford, its inline with other teasers
- the "by [company name]" prefix is darker, to pass WCAG 2.0
contrast guidelines and meet DAC accreditation

These updates must be released to v4 and v5, as key pages on
ft.com is one major behind the app.