Do not apply typographic styles using placeholders.
Do not include typography styles within the placeholder, as `@extend`
interleaves selectors unnecessarily for the progressive font fallback.
This may cause large selectors.

E.g. this is ok, if a `.loading` class is on the a parent element
(i.e. the html element) then we'll get type fallback styles:

```scss
%type {
    content: 'type';
    @at-root .loading & {
        content: 'type fallback';
    }
}
p {
    @extend %type;
}
```

```css
p {
  content: "type";
}
.loading p {
  content: "type fallback";
}
```

But this is not ok, we're creating too many selectors, `@extend` is interleaving them:

```scss
%type {
    content: 'type';
    @at-root .loading & {
        content: 'type fallback';
    }
}
.container p {
    @extend %type;
}
```

```css
.container p {
  content: "type";
}
.loading .container p, .container .loading p {
  content: "type fallback";
}
```